% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{rel_diff}
\alias{rel_diff}
\alias{abs_diff}
\title{Absolute and relative difference}
\usage{
rel_diff(x, y)

abs_diff(x, y)
}
\arguments{
\item{x}{A \link{double} vector.}

\item{y}{A \link{double} vector.}
}
\value{
A numeric vector.
}
\description{
Calculate absolute differences with \code{abs_diff()} and
relative differences with \code{rel_diff()}
}
