% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_threads}
\alias{get_threads}
\alias{set_threads}
\title{Get and set threads for parallel computation}
\usage{
get_threads()

set_threads(n)
}
\arguments{
\item{n}{\verb{[integer(1)]} - Number of threads to use package wide.}
}
\value{
Gets or set the number of specified threads.
}
\description{
Get and set threads for parallel computation
}
\examples{

library(cppdoubles)

get_threads() # The default = max_threads / 4

set_threads(1)
get_threads() # Should now be 1

}
