% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\alias{get_tolerance}
\alias{set_tolerance}
\title{Get and set package-wide tolerance}
\usage{
get_tolerance()

set_tolerance(x)
}
\arguments{
\item{x}{\verb{[double(1)]} - Tolerance to be used across all
cppdoubles functions.}
}
\value{
Either sets or gets the tolerance to be used package-wide.
}
\description{
Get and set package-wide tolerance
}
